Don't clone mainline pads
Inspired by some discussion with TimToady++ , this is an attempt to make END
more useful, but it also seems to help with optimization prospects a great
deal as well.  A pad which belongs to a mainline (defined recursively below)
will not be deeply cloned; any variable in it at runtime aliases the same
variable at compile time.

UNIT is a mainline.
Any statement-level bare block contained in a mainline is itself a mainline.
Any package definition in a mainline is itself a mainline.