Redesign to cross-domain communication system
Instead of using a custom interface which is compiled once and shared
between versions of Kernel.dll, a system which proved fragile and
inflexible, we now coopt System.Collections.IDictionary.

Also adds a C# code module to the compiler, finally allowing all
aspects of the compiler/compilee interface to be controlled by current
code.