Add a max_buffer_size to the ObjectPipe
The blocking aspect should work both ways, not only the reader should block on the writer
but it is also important that the writer blocks on the reader so that we only advance a
reasonable set of data prior to any consumption. Just like unix pipes.

Added an example code that demonstrates that.