make the example more explicit
The semantics of batch-processing imply that we may actually not use all the data
that we produce. But we dont wait for each value to be used to produce another one
The output of the example shows that the thread will produce a set of objects
before the context switches again.