First attempt to use a per-user directory for the module cache
Data is stored in Local Application Data\NieczaModuleCache, which Mono will
map to $HOME/.local/share/NieczaModuleCache.  The --obj-dir object actually
does something now (how long has it been broken?) and can be used to isolate
multiple instances of Niecza, and is used internally to make bootstrapping
work.

One known wart is that $*PERL<version> may become out of date because it is
only updated when the setting needs to be recompiled, and the build system
no longer always forces a recompilation of the user-wide setting.